# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "LoL Analytics API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/matches/{match_id}": {
   "get": {
    "tags": [
     "matches"
    ],
    "summary": "Get Match",
    "description": "The full scoreboard for a match we already hold.",
    "operationId": "get_match_api_matches__match_id__get",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 6,
       "maxLength": 32,
       "description": "Riot match id, which carries its own platform: EUW1_7986353741.",
       "title": "Match Id"
      },
      "description": "Riot match id, which carries its own platform: EUW1_7986353741."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MatchDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/matches/{match_id}/story": {
   "get": {
    "tags": [
     "matches"
    ],
    "summary": "Get Story",
    "description": "Each team's chance to win over the game, the moments that decided it,\nand every player's deaths and takedowns weighed.\n\nFetches the game's timeline from Riot if we do not hold it: one call, once.",
    "operationId": "get_story_api_matches__match_id__story_get",
    "parameters": [
     {
      "name": "match_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 6,
       "maxLength": 32,
       "description": "Riot match id, which carries its own platform: EUW1_7986353741.",
       "title": "Match Id"
      },
      "description": "Riot match id, which carries its own platform: EUW1_7986353741."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/GameStoryResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "BadgeOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "detail": {
      "type": "string",
      "title": "Detail"
     }
    },
    "type": "object",
    "required": [
     "id",
     "label",
     "detail"
    ],
    "title": "BadgeOut",
    "description": "A badge with the rule that earned it and this player's own figure.\n\nThe detail is composed per request rather than stored, so the wording can\nimprove without rescoring the corpus, and a badge can never appear with an\nexplanation that no longer matches the rule that awarded it."
   },
...
   "ChampionRef": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "icon_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Icon Url"
     },
     "slug": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Slug",
      "readOnly": true
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "slug"
    ],
    "title": "ChampionRef"
   },
...
   "CurvePoint": {
    "properties": {
     "ms": {
      "type": "integer",
      "title": "Ms"
     },
     "blue": {
      "type": "number",
      "title": "Blue"
     },
     "frame": {
      "type": "boolean",
      "title": "Frame",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "ms",
     "blue"
    ],
    "title": "CurvePoint"
   },
...
   "DeathOut": {
    "properties": {
     "ms": {
      "type": "integer",
      "title": "Ms"
     },
     "killer": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StoryPlayerRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "assisters": {
      "type": "integer",
      "title": "Assisters",
      "default": 0
     },
     "x": {
      "type": "integer",
      "title": "X",
      "default": 0
     },
     "y": {
      "type": "integer",
      "title": "Y",
      "default": 0
     },
     "traded": {
      "type": "boolean",
      "title": "Traded"
     },
     "cost": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost"
     }
    },
    "type": "object",
    "required": [
     "ms",
     "traded"
    ],
    "title": "DeathOut"
   },
...
   "GameStoryResponse": {
    "properties": {
     "match_id": {
      "type": "string",
      "title": "Match Id"
     },
     "available": {
      "type": "boolean",
      "title": "Available"
     },
     "pending": {
      "type": "boolean",
      "title": "Pending",
      "default": false
     },
     "retry_after": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Retry After"
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "queue_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Queue Id"
     },
     "duration_ms": {
      "type": "integer",
      "title": "Duration Ms",
      "default": 0
     },
     "blue_won": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Blue Won"
     },
     "curve": {
      "items": {
       "$ref": "#/components/schemas/CurvePoint"
      },
      "type": "array",
      "title": "Curve"
     },
     "moments": {
      "items": {
       "$ref": "#/components/schemas/MomentOut"
      },
      "type": "array",
      "title": "Moments"
     },
     "players": {
      "items": {
       "$ref": "#/components/schemas/PlayerStoryOut"
      },
      "type": "array",
      "title": "Players"
     },
     "model": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/StoryModelOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "map_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Map Url"
     }
    },
    "type": "object",
    "required": [
     "match_id",
     "available"
    ],
    "title": "GameStoryResponse"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ItemRef": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "icon_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Icon Url"
     },
     "slug": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Slug",
      "readOnly": true
     }
    },
    "type": "object",
    "required": [
     "id",
     "slug"
    ],
    "title": "ItemRef"
   },
...
   "MatchDetailResponse": {
    "properties": {
     "match_id": {
      "type": "string",
      "title": "Match Id"
     },
     "queue_id": {
      "type": "integer",
      "title": "Queue Id"
     },
     "queue_name": {
      "type": "string",
      "title": "Queue Name"
     },
     "patch": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Patch"
     },
     "game_creation": {
      "type": "integer",
      "title": "Game Creation"
     },
     "game_duration": {
      "type": "integer",
      "title": "Game Duration"
     },
     "is_remake": {
      "type": "boolean",
      "title": "Is Remake",
      "default": false
     },
     "platform": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Platform"
     },
     "score_withheld": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score Withheld"
     },
     "teams": {
      "items": {
       "items": {
        "$ref": "#/components/schemas/ScoreboardPlayer"
       },
       "type": "array"
      },
      "type": "array",
      "title": "Teams"
     },
     "objectives": {
      "items": {
       "$ref": "#/components/schemas/TeamObjectives"
      },
      "type": "array",
      "title": "Objectives"
     },
     "model": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ScoreModelOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "match_id",
     "queue_id",
     "queue_name",
     "game_creation",
     "game_duration"
    ],
    "title": "MatchDetailResponse"
   },
...
   "MomentOut": {
    "properties": {
     "start_ms": {
      "type": "integer",
      "title": "Start Ms"
     },
     "end_ms": {
      "type": "integer",
      "title": "End Ms"
     },
     "swing": {
      "type": "number",
      "title": "Swing"
     },
     "team": {
      "type": "integer",
      "title": "Team"
     },
     "text": {
      "type": "string",
      "title": "Text"
     }
    },
    "type": "object",
    "required": [
     "start_ms",
     "end_ms",
     "swing",
     "team",
     "text"
    ],
    "title": "MomentOut"
   },
...
   "PlayerStoryOut": {
    "properties": {
     "participant_index": {
      "type": "integer",
      "title": "Participant Index"
     },
     "puuid": {
      "type": "string",
      "title": "Puuid"
     },
     "team_id": {
      "type": "integer",
      "title": "Team Id"
     },
     "riot_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Riot Id"
     },
     "champion": {
      "$ref": "#/components/schemas/ChampionRef"
     },
     "position": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Position"
     },
     "deaths": {
      "type": "integer",
      "title": "Deaths",
      "default": 0
     },
     "untraded": {
      "type": "integer",
      "title": "Untraded",
      "default": 0
     },
     "win_lost": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Win Lost"
     },
     "takedowns": {
      "type": "integer",
      "title": "Takedowns",
      "default": 0
     },
     "converted": {
      "type": "integer",
      "title": "Converted",
      "default": 0
     },
     "win_gained": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Win Gained"
     },
     "contests": {
      "type": "integer",
      "title": "Contests",
      "default": 0
     },
     "contests_won": {
      "type": "integer",
      "title": "Contests Won",
      "default": 0
     },
     "death_list": {
      "items": {
       "$ref": "#/components/schemas/DeathOut"
      },
      "type": "array",
      "title": "Death List"
     },
     "takedown_list": {
      "items": {
       "$ref": "#/components/schemas/TakedownOut"
      },
      "type": "array",
      "title": "Takedown List"
     }
    },
    "type": "object",
    "required": [
     "participant_index",
     "puuid",
     "team_id",
     "champion"
    ],
    "title": "PlayerStoryOut"
   },
...
   "PublishedComponentOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "measures": {
      "type": "string",
      "title": "Measures"
     }
    },
    "type": "object",
    "required": [
     "id",
     "label",
     "measures"
    ],
    "title": "PublishedComponentOut",
    "description": "One of the things the score is made of, as the method page lists them."
   },
...
   "RuneRef": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "icon_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Icon Url"
     }
    },
    "type": "object",
    "title": "RuneRef"
   },
...
   "ScoreModelOut": {
    "properties": {
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "components": {
      "items": {
       "$ref": "#/components/schemas/PublishedComponentOut"
      },
      "type": "array",
      "title": "Components"
     },
     "weights": {
      "additionalProperties": {
       "additionalProperties": {
        "type": "number"
       },
       "type": "object"
      },
      "type": "object",
      "title": "Weights"
     },
     "samples": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Samples"
     },
     "note": {
      "type": "string",
      "title": "Note"
     }
    },
    "type": "object",
    "required": [
     "version",
     "components",
     "weights",
     "note"
    ],
    "title": "ScoreModelOut",
    "description": "How the Riftline score is computed, published with every scoreboard.\n\nitero exposes its draft model and it is the most trustworthy thing on their\nsite. A rating that cannot be taken apart is a rating that has to be taken\non faith, which is the opposite of what this whole project is for."
   },
   "ScoreboardPlayer": {
    "properties": {
     "puuid": {
      "type": "string",
      "title": "Puuid"
     },
     "riot_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Riot Id"
     },
     "champion": {
      "$ref": "#/components/schemas/ChampionRef"
     },
     "team_id": {
      "type": "integer",
      "title": "Team Id"
     },
     "position": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Position"
     },
     "win": {
      "type": "boolean",
      "title": "Win",
      "default": false
     },
     "champ_level": {
      "type": "integer",
      "title": "Champ Level",
      "default": 0
     },
     "laning_score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Laning Score"
     },
     "laning_label": {
      "anyOf": [
       {
        "type": "string",
        "enum": [
         "won_big",
         "won",
         "even",
         "lost",
         "lost_big"
        ]
       },
       {
        "type": "null"
       }
      ],
      "title": "Laning Label"
     },
     "kills": {
      "type": "integer",
      "title": "Kills",
      "default": 0
     },
     "deaths": {
      "type": "integer",
      "title": "Deaths",
      "default": 0
     },
     "assists": {
      "type": "integer",
      "title": "Assists",
      "default": 0
     },
     "kill_participation": {
      "type": "number",
      "title": "Kill Participation",
      "default": 0.0
     },
     "cs": {
      "type": "integer",
      "title": "Cs",
      "default": 0
     },
     "cs_per_min": {
      "type": "number",
      "title": "Cs Per Min",
      "default": 0.0
     },
     "gold_earned": {
      "type": "integer",
      "title": "Gold Earned",
      "default": 0
     },
     "damage_to_champions": {
      "type": "integer",
      "title": "Damage To Champions",
      "default": 0
     },
     "damage_taken": {
      "type": "integer",
      "title": "Damage Taken",
      "default": 0
     },
     "vision_score": {
      "type": "integer",
      "title": "Vision Score",
      "default": 0
     },
     "wards_placed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wards Placed"
     },
     "wards_killed": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wards Killed"
     },
     "control_wards": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Control Wards"
     },
     "items": {
      "items": {
       "$ref": "#/components/schemas/ItemRef"
      },
      "type": "array",
      "title": "Items"
     },
     "trinket": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ItemRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "spells": {
      "items": {
       "$ref": "#/components/schemas/SpellRef"
      },
      "type": "array",
      "title": "Spells"
     },
     "keystone": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/RuneRef"
       },
       {
        "type": "null"
       }
      ]
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     },
     "placement": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Placement"
     },
     "badges": {
      "items": {
       "$ref": "#/components/schemas/BadgeOut"
      },
      "type": "array",
      "title": "Badges"
     }
    },
    "type": "object",
    "required": [
     "puuid",
     "champion",
     "team_id"
    ],
    "title": "ScoreboardPlayer",
    "description": "One player on the expanded scoreboard.\n\nEverything the row cannot fit. Served from its own endpoint rather than in\nthe history payload: twenty matches times ten players times this many fields\nis a page nobody asked for on every scroll."
   },
...
   "SpellRef": {
    "properties": {
     "id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Id"
     },
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "icon_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Icon Url"
     }
    },
    "type": "object",
    "title": "SpellRef"
   },
...
   "StoryModelOut": {
    "properties": {
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "published": {
      "type": "boolean",
      "title": "Published"
     },
     "withheld": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Withheld"
     },
     "trained_games": {
      "type": "integer",
      "title": "Trained Games",
      "default": 0
     },
     "trained_queue": {
      "type": "integer",
      "title": "Trained Queue",
      "default": 420
     },
     "accuracy": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Accuracy"
     },
     "phases": {
      "items": {
       "$ref": "#/components/schemas/StoryPhaseOut"
      },
      "type": "array",
      "title": "Phases"
     }
    },
    "type": "object",
    "required": [
     "version",
     "published"
    ],
    "title": "StoryModelOut"
   },
...
   "TeamObjectives": {
    "properties": {
     "team_id": {
      "type": "integer",
      "title": "Team Id"
     },
     "win": {
      "type": "boolean",
      "title": "Win",
      "default": false
     },
     "kills": {
      "type": "integer",
      "title": "Kills",
      "default": 0
     },
     "gold": {
      "type": "integer",
      "title": "Gold",
      "default": 0
     },
     "objectives_known": {
      "type": "boolean",
      "title": "Objectives Known",
      "default": true
     },
     "baron": {
      "type": "integer",
      "title": "Baron",
      "default": 0
     },
     "dragon": {
      "type": "integer",
      "title": "Dragon",
      "default": 0
     },
     "herald": {
      "type": "integer",
      "title": "Herald",
      "default": 0
     },
     "tower": {
      "type": "integer",
      "title": "Tower",
      "default": 0
     },
     "inhibitor": {
      "type": "integer",
      "title": "Inhibitor",
      "default": 0
     },
     "bans": {
      "items": {
       "$ref": "#/components/schemas/ChampionRef"
      },
      "type": "array",
      "title": "Bans"
     }
    },
    "type": "object",
    "required": [
     "team_id"
    ],
    "title": "TeamObjectives",
    "description": "What one side did. Kills and gold are summed from its own players.\n\n`objectives_known` is false when Riot's team objects do not line up with the\nsides on the scoreboard, which is the case in Arena. Then the structure and\nmonster counts below are all zero because they are unknown, not because\nnothing was taken, and the UI omits them."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}